modularizando la data
