+++ src/utils/Data.js
@@ -1,0 +1,45 @@
+import ImgBot1 from "../assets/img/desktop/image-curiosity.jpg"
+import ImgBot2 from "../assets/img/desktop/image-deep-earth.jpg"
+import ImgBot3 from "../assets/img/desktop/image-fisheye.jpg"
+import ImgBot4 from "../assets/img/desktop/image-from-above.jpg"
+import ImgBot5 from "../assets/img/desktop/image-grid.jpg"
+import ImgBot6 from "../assets/img/desktop/image-pocket-borealis.jpg"
+import ImgBot7 from "../assets/img/desktop/image-night-arcade.jpg"
+import ImgBot8 from "../assets/img/desktop/image-soccer-team.jpg"
+
+export const items = ["about", "carrers", "events", "products", "support"];
+
+export const imgCards = [
+    {
+        imagenes: ImgBot1,
+        title: "DEEP EARTH"
+    },
+    {
+        imagenes: ImgBot2,
+        title: "NIGHT ARCADE"
+    },
+    {
+        imagenes: ImgBot3,
+        title: "SOCCER TEAM VR"
+    },
+    {
+        imagenes: ImgBot4,
+        title: "THE GRID"
+    },
+    {
+        imagenes: ImgBot5,
+        title: "FROM UP ABOVE VR"
+    },
+    {
+        imagenes: ImgBot6,
+        title: "POCKET BOREALIS"
+    },
+    {
+        imagenes: ImgBot7,
+        title: "THE CURIOSITY"
+    },
+    {
+        imagenes: ImgBot8,
+        title: "MAKE IT FISHEYE"
+    }
+]
